requirements-test.txt
-r web/requirements.txt
pytest==8.3.4
pytest-flask==1.3.0
